Add files via upload
In this exercise we will run an e-commerce system. We will keep its name and the system of users for the system.
For each user we will save his username, login password and residential address. Any user
Is either a seller or a buyer (there is no one who is both a seller and a buyer)
All the feedback he received from the customers, so that each feedback contains information about the giver of the feedback, the date of delivery
The feedback and verbal assessment. Also, keep all the products he sells for him. There is no need
Manage inventory. For each buyer we will keep the shopping cart containing the products the customer has chosen
And not yet ordered. Each product is associated with a particular category (children, electricity, office and clothing), it has a name, a price
And a unique serial number in the system) The same product at different sellers will have a different serial number (
When a customer places an order, he can only choose products that are in his shopping cart. for
The order will keep the product items, the total order price, the seller details and details
The buyer. When paying for the order the items will be removed from the buyer's shopping cart.